+++ CHANGELOG.md
@@ -8,3 +8,7 @@
 - Added `trailingAnimation` property which is a `Animatable<double>`
 - Added `trailingAnimationBuilder` property to customize the trailing widget animation
 - Added `trailingAnimationEnabled` property which enabled/disabled trailing animation is true by default
+
+## 0.1.1
+- Bug fix for initialExpandedIndex/initialExpandedIndexes `trailingAnimation` not working
+- Added a better description for the package